+++ neoforge/src/main/java/io/github/mortuusars/monobank/neoforge/ItemHandlerCapabilityBlockEntity.java
@@ -1,0 +1,7 @@
+package io.github.mortuusars.monobank.neoforge;
+
+import net.neoforged.neoforge.items.IItemHandler;
+
+public interface ItemHandlerCapabilityBlockEntity {
+    IItemHandler monobank$getItemHandlerCapability();
+}