+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## UNRELEASED
 - Updated to 1.21.1
+- Updated Monobank textures.
 - When Monobank is unlocked with combination, items used as keys will now return straight to player's inventory instead of dropping on the ground.
 - Added Lantern, Shears, Rail, Flower Pot and Stone Button to village combination loot tables. 
 - Restructured config.
   - Added several config options related to locks and combination.
 - Removed all thief-related functionality in favor of integration with `Thief` mod.
-- Monobanks are no longer carryable with `Carry On` mod.     
+- Monobanks are no longer carryable with `Carry On` mod.
+- Create's **Threshold Switch** will now show correct amounts for Monobank. 
 - Fixed 'Break In Attempted' or 'Break In Successful' warnings being resettable by a player that does not own the Monobank.
 - Break In warnings no longer showing for a player that does not own the Monobank.
 - Break In warnings now clear after 3 shows (UI opened 3 times).